Removes deprecated use of :confirm in scaffold templates
The :confirm option has been deprecated and will be removed in Rails
4.1.

This commit changes all instances of: ":confirm => 'Are you sure?'"
with ":data => { confirm: 'Are you sure?' }" in the scaffold templates.

The exact warning in is:
DEPRECATION WARNING: :confirm option is deprecated and will be removed
from Rails 4.1. Use 'data: { confirm: 'Text' }' instead.