Remove `content_tag_for` for Rails 5 compatibility
`content_tag_for` was extracted to a separate gem for Rails 5. This no longer depends on that functionality.

This fixes #43. 

I've tested this in Rails 5, but not other versions. The cloned tests are throwing errors about missing bootstrap files, so I can't say much about the tests for this other than "it works for my project."